+++ tests/src/test/java/com/regnosys/ingest/dtcc/DtccIngestion9ServiceTest.java
@@ -1,14 +1,17 @@
-package com.regnosys.granite.ingestor.dtcc;
+package com.regnosys.ingest.dtcc;
 
 import cdm.event.workflow.WorkflowStep;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import com.google.common.collect.ImmutableList;
 import com.google.common.io.Resources;
+import com.regnosys.ingest.test.framework.ingestor.IngestionReport;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTest;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTestUtil;
 import com.regnosys.ingest.test.framework.ingestor.service.IngestionFactory;
 import com.regnosys.ingest.test.framework.ingestor.service.IngestionService;
 import com.regnosys.ingest.test.framework.ingestor.synonym.MappingReport;
 import com.regnosys.ingest.test.framework.ingestor.synonym.MappingResult;
+import com.regnosys.ingest.test.framework.ingestor.testing.Expectation;
 import org.finos.cdm.CdmRuntimeModule;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.provider.Arguments;
@@ -19,24 +22,29 @@ import java.net.URL;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
-class DtccIngestion11ServiceTest  extends IngestionTest<WorkflowStep> {
-    private static final Logger LOGGER = LoggerFactory.getLogger(DtccIngestion11ServiceTest.class);
+class DtccIngestion9ServiceTest extends IngestionTest<WorkflowStep> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(DtccIngestion9ServiceTest.class);
 
-	private static final String DTCC_11_0_FILES_DIR = "cdm-sample-files/dtcc-11-0/";
+	private static final String DTCC_9_0_FILES_DIR = "cdm-sample-files/dtcc-9-0/";
 
 	private static ImmutableList<URL> EXPECTATION_FILES = ImmutableList.<URL>builder()
-			.add(Resources.getResource(DTCC_11_0_FILES_DIR + "expectations.json"))
+			.add(Resources.getResource(DTCC_9_0_FILES_DIR + "expectations.json"))
 			.build();
 
-	private static IngestionService dtcc11IngestionService;
+	private static IngestionService dtcc9IngestionService;
 
 	@BeforeAll
 	static void setup() {
 		CdmRuntimeModule runtimeModule = new CdmRuntimeModule();
 		initialiseIngestionFactory(runtimeModule, IngestionTestUtil.getPostProcessors(runtimeModule));
-		dtcc11IngestionService = IngestionFactory.getInstance().getDtcc11();
+		dtcc9IngestionService = IngestionFactory.getInstance().getDtcc9();
 	}
-	
+
+	@Override
+	protected void assertExpectations(Expectation expectation, IngestionReport<WorkflowStep> ingested) throws JsonProcessingException {
+
+	}
+
 	@Override
 	protected Class<WorkflowStep> getClazz() {
 		return WorkflowStep.class;
@@ -44,7 +52,7 @@ class DtccIngestion11ServiceTest  extends IngestionTest<WorkflowStep> {
 
 	@Override
 	protected IngestionService ingestionService() {
-		return dtcc11IngestionService;
+		return dtcc9IngestionService;
 	}
 
 	@SuppressWarnings("unused")//used by the junit parameterized test
