+++ tests/src/test/java/com/regnosys/ingest/fpml/Fpml510ProcessesIngestionServiceTest.java
@@ -1,11 +1,10 @@
-package com.regnosys.granite.ingestor.cme;
+package com.regnosys.ingest.fpml;
 
 import cdm.event.workflow.WorkflowStep;
 import com.google.common.collect.ImmutableList;
 import com.google.common.io.Resources;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTest;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTestUtil;
-import com.regnosys.ingest.test.framework.ingestor.service.IngestionFactory;
 import com.regnosys.ingest.test.framework.ingestor.service.IngestionService;
 import org.finos.cdm.CdmRuntimeModule;
 import org.junit.jupiter.api.BeforeAll;
@@ -14,21 +13,23 @@ import org.junit.jupiter.params.provider.Arguments;
 import java.net.URL;
 import java.util.stream.Stream;
 
-class CMESubmissionTest extends IngestionTest<WorkflowStep>{
+import static org.isda.cdm.util.IngestionEnvUtil.getFpml5ConfirmationToWorkflowStep;
 
-	private static final String CME_SUBMISSION_1_17_FILES_DIR = "cdm-sample-files/cme-submission-irs-1-0/";
+public class Fpml510ProcessesIngestionServiceTest extends IngestionTest<WorkflowStep> {
+
+	private static final String PROCESSES = "cdm-sample-files/fpml-5-10/processes/";
 
 	private static ImmutableList<URL> EXPECTATION_FILES = ImmutableList.<URL>builder()
-			.add(Resources.getResource(CME_SUBMISSION_1_17_FILES_DIR + "expectations.json"))
+			.add(Resources.getResource(PROCESSES + "expectations.json"))
 			.build();
-	
+
 	private static IngestionService ingestionService;
 
 	@BeforeAll
 	static void setup() {
 		CdmRuntimeModule runtimeModule = new CdmRuntimeModule();
 		initialiseIngestionFactory(runtimeModule, IngestionTestUtil.getPostProcessors(runtimeModule));
-		ingestionService = IngestionFactory.getInstance().getCmeSubmissionIrs1();
+		ingestionService = getFpml5ConfirmationToWorkflowStep();
 	}
 	
 	@Override
@@ -45,5 +46,4 @@ class CMESubmissionTest extends IngestionTest<WorkflowStep>{
 	private static Stream<Arguments> fpMLFiles() {
 		return readExpectationsFrom(EXPECTATION_FILES);
 	}
-
 }
