+++ tests/src/test/java/com/regnosys/ingest/fpml/Fpml510ProductIngestionServiceTest.java
@@ -1,8 +1,6 @@
-package com.regnosys.granite.ingestor.fpml;
+package com.regnosys.ingest.fpml;
 
-import cdm.event.workflow.WorkflowStep;
-import com.google.common.collect.ImmutableList;
-import com.google.common.io.Resources;
+import cdm.event.common.TradeState;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTest;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTestUtil;
 import com.regnosys.ingest.test.framework.ingestor.service.IngestionService;
@@ -10,18 +8,13 @@ import org.finos.cdm.CdmRuntimeModule;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.provider.Arguments;
 
-import java.net.URL;
 import java.util.stream.Stream;
 
-import static org.isda.cdm.util.IngestionEnvUtil.getFpml5ConfirmationToWorkflowStep;
+import static org.isda.cdm.util.IngestionEnvUtil.getFpml5ConfirmationToTradeState;
 
-public class Fpml510IncompleteProcessesIngestionServiceTest extends IngestionTest<WorkflowStep> {
+public class Fpml510ProductIngestionServiceTest extends IngestionTest<TradeState> {
 
-	private static final String INCOMPLETE_PROCESSES = "cdm-sample-files/fpml-5-10/incomplete-processes/";
-
-	private static ImmutableList<URL> EXPECTATION_FILES = ImmutableList.<URL>builder()
-			.add(Resources.getResource(INCOMPLETE_PROCESSES + "expectations.json"))
-			.build();
+	private static final String BASE_DIR = "cdm-sample-files/fpml-5-10/products/";
 
 	private static IngestionService ingestionService;
 
@@ -29,12 +22,12 @@ public class Fpml510IncompleteProcessesIngestionServiceTest extends IngestionTes
 	static void setup() {
 		CdmRuntimeModule runtimeModule = new CdmRuntimeModule();
 		initialiseIngestionFactory(runtimeModule, IngestionTestUtil.getPostProcessors(runtimeModule));
-		ingestionService = getFpml5ConfirmationToWorkflowStep();
+		ingestionService = getFpml5ConfirmationToTradeState();
 	}
 	
 	@Override
-	protected Class<WorkflowStep> getClazz() {
-		return WorkflowStep.class;
+	protected Class<TradeState> getClazz() {
+		return TradeState.class;
 	}
 
 	@Override
@@ -44,6 +37,6 @@ public class Fpml510IncompleteProcessesIngestionServiceTest extends IngestionTes
 
 	@SuppressWarnings("unused")//used by the junit parameterized test
 	private static Stream<Arguments> fpMLFiles() {
-		return readExpectationsFrom(EXPECTATION_FILES);
+		return readExpectationsFromPath(BASE_DIR);
 	}
 }
