+++ tests/src/test/java/com/regnosys/ingest/fpml/NativeCdmEventsIngestionServiceTest.java
@@ -1,6 +1,8 @@
-package com.regnosys.granite.ingestor.fpml;
+package com.regnosys.ingest.fpml;
 
-import cdm.event.common.TradeState;
+import cdm.event.workflow.WorkflowStep;
+import com.google.common.collect.ImmutableList;
+import com.google.common.io.Resources;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTest;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTestUtil;
 import com.regnosys.ingest.test.framework.ingestor.service.IngestionService;
@@ -8,13 +10,18 @@ import org.finos.cdm.CdmRuntimeModule;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.params.provider.Arguments;
 
+import java.net.URL;
 import java.util.stream.Stream;
 
-import static org.isda.cdm.util.IngestionEnvUtil.getFpml5ConfirmationToTradeState;
+import static org.isda.cdm.util.IngestionEnvUtil.getFpml5ConfirmationToWorkflowStep;
 
-public class Fpml510IncompleteProductIngestionServiceTest extends IngestionTest<TradeState> {
+public class NativeCdmEventsIngestionServiceTest extends IngestionTest<WorkflowStep> {
 
-	private static final String INCOMPLETE_BASE = "cdm-sample-files/fpml-5-10/incomplete-products/";
+	private static final String BASE_DIR = "cdm-sample-files/native-cdm-events/";
+
+	private static ImmutableList<URL> EXPECTATION_FILES = ImmutableList.<URL>builder()
+		.add(Resources.getResource(BASE_DIR + "expectations.json"))
+		.build();
 
 	private static IngestionService ingestionService;
 
@@ -22,12 +29,12 @@ public class Fpml510IncompleteProductIngestionServiceTest extends IngestionTest<
 	static void setup() {
 		CdmRuntimeModule runtimeModule = new CdmRuntimeModule();
 		initialiseIngestionFactory(runtimeModule, IngestionTestUtil.getPostProcessors(runtimeModule));
-		ingestionService = getFpml5ConfirmationToTradeState();
+		ingestionService = getFpml5ConfirmationToWorkflowStep();
 	}
 
 	@Override
-	protected Class<TradeState> getClazz() {
-		return TradeState.class;
+	protected Class<WorkflowStep> getClazz() {
+		return WorkflowStep.class;
 	}
 
 	@Override
@@ -37,6 +44,6 @@ public class Fpml510IncompleteProductIngestionServiceTest extends IngestionTest<
 
 	@SuppressWarnings("unused")//used by the junit parameterized test
 	private static Stream<Arguments> fpMLFiles() {
-		return readExpectationsFromPath(INCOMPLETE_BASE);
+		return readExpectationsFrom(EXPECTATION_FILES);
 	}
 }
