+++ tests/src/test/java/com/regnosys/ingest/ore/OreTradeTest.java
@@ -1,10 +1,11 @@
-package com.regnosys.granite.ingestor.fpml;
+package com.regnosys.ingest.ore;
 
-import cdm.event.workflow.WorkflowStep;
+import cdm.event.common.TradeState;
 import com.google.common.collect.ImmutableList;
 import com.google.common.io.Resources;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTest;
 import com.regnosys.ingest.test.framework.ingestor.IngestionTestUtil;
+import com.regnosys.ingest.test.framework.ingestor.service.IngestionFactory;
 import com.regnosys.ingest.test.framework.ingestor.service.IngestionService;
 import org.finos.cdm.CdmRuntimeModule;
 import org.junit.jupiter.api.BeforeAll;
@@ -13,28 +14,26 @@ import org.junit.jupiter.params.provider.Arguments;
 import java.net.URL;
 import java.util.stream.Stream;
 
-import static org.isda.cdm.util.IngestionEnvUtil.getFpml5ConfirmationToWorkflowStep;
+class OreTradeTest extends IngestionTest<TradeState>{
 
-public class Fpml510ProcessesIngestionServiceTest extends IngestionTest<WorkflowStep> {
-
-	private static final String PROCESSES = "cdm-sample-files/fpml-5-10/processes/";
+	private static final String ORE_1_0_39_FILES_DIR = "cdm-sample-files/ore-1-0-39/";
 
 	private static ImmutableList<URL> EXPECTATION_FILES = ImmutableList.<URL>builder()
-			.add(Resources.getResource(PROCESSES + "expectations.json"))
+			.add(Resources.getResource(ORE_1_0_39_FILES_DIR + "expectations.json"))
 			.build();
-
+	
 	private static IngestionService ingestionService;
 
 	@BeforeAll
 	static void setup() {
 		CdmRuntimeModule runtimeModule = new CdmRuntimeModule();
 		initialiseIngestionFactory(runtimeModule, IngestionTestUtil.getPostProcessors(runtimeModule));
-		ingestionService = getFpml5ConfirmationToWorkflowStep();
+		ingestionService = IngestionFactory.getInstance().getOre1039();
 	}
 	
 	@Override
-	protected Class<WorkflowStep> getClazz() {
-		return WorkflowStep.class;
+	protected Class<TradeState> getClazz() {
+		return TradeState.class;
 	}
 
 	@Override
@@ -46,4 +45,5 @@ public class Fpml510ProcessesIngestionServiceTest extends IngestionTest<Workflow
 	private static Stream<Arguments> fpMLFiles() {
 		return readExpectationsFrom(EXPECTATION_FILES);
 	}
+
 }
